requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4; sys_platform == 'darwin'  # macOS specific
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
executing==2.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.1
langgraph-sdk==0.1.33
langsmith==0.1.137
marshmallow==3.23.0
matplotlib-inline==0.1.7
MouseInfo==0.1.3; sys_platform == 'darwin'  # macOS specific
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.52.1
orjson==3.10.10
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54; sys_platform == 'darwin'  # macOS specific
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyGetWindow==0.0.9; sys_platform == 'darwin'  # macOS specific
Pygments==2.18.0
PyMsgBox==1.0.9; sys_platform == 'darwin'  # macOS specific
pyobjc-core==10.3.1; sys_platform == 'darwin'  # macOS specific
pyobjc-framework-Cocoa==10.3.1; sys_platform == 'darwin'  # macOS specific
pyobjc-framework-Quartz==10.3.1; sys_platform == 'darwin'  # macOS specific
pyperclip==1.9.0
PyRect==0.2.0; sys_platform == 'darwin'  # macOS specific
PyScreeze==1.0.1; sys_platform == 'darwin'  # macOS specific
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rubicon-objc==0.4.9; sys_platform == 'darwin'  # macOS specific
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.16.0